covid_tracker/requirements.txt
asgiref==3.3.4
Django==3.2.4
django-mysql==3.12.0
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
phonenumbers==8.12.26
pytz==2021.1
sqlparse==0.4.1

